Make subscribers array for AWS S3 storage
