+++ subscriber.html
@@ -14,6 +14,6 @@
         <button type="submit" id="testButton">Add subscriber</button>
       </form>
 
-<script type="module" src="Assets/subscribers.js"></script>
+<script type="module" src="Assets/subscribe.js"></script>
 </body>
 </html>
